segmentationRoutines.py
edits to put the temp hdf5 file in a temporary working directory, and
set the cwd argument to popen() appropriately.

this resolves a possible issue where one might not have read access to
the current directory (like off a cd) or where the local directory is
actually remote (slow access)

furthermore, it keeps the current directory clean of temp files, which
is good.