scanimage routine changes
lots of little changes to make the routines more robust.

pulled out automatic alignment and calculating the epoch averages --
those are now in the import notebook

infact, the entire import code for "readMultiOdorEpoch" is now in the
import note book-- this function is officially deprecated.

note that 'addTrialToEpoch' is still used (no alignFlag tho)