added stimulator import to parseXSG()
Stimulator field now holds a dictionary with key/values that are output
name / array

Note that I've put the ephys stimulations in here as well- a bit
muddled but I think it'll work fine.  it's hard coded for a single
ephys channel-- in the future we'll need to change this if we dual
patch but for now it's fine.

Tested with simple square pulses only--- additive pulses will require
more logic.