baseline_splines() / mask_deviations()
made some changes here to drop the order of spline interpolation if the
number of control points is 3 or less (points must be > order, always).

added some logic to deal with the end control points being undefined
(duplicated the next one in)

added ability to pass through std_cutoff from baseline splines

all of these changes are due to it being harder to correct the baseline
when the stimulus is long relative to the whole trace.  gonna take some
fiddling and could be improved, still.