misc changes to cellPicker
biggest change is making sure that the contours and fits in the info
panel match what is used for picking cells.  needed to change to the
thresholded NMF, that was wicked confusing.

changed modes back down to 4 and am using all modes instead of ignoring
the first one.  i think our heuristics are good enough we can include
that mode even though it's basically a DC.

changed the cell-like heuristic to do the ABS of fit width, as I
realized that it is sign invariant.  also changed the initial values in
the fit to be at least 1--- doesn't really seem to make a difference
but when those values are 0 it's indicative of a poor fit.  maybe a bit
better now.

also a misc. change in here for the diskSize, to match the default
label in the GUI