adding label mode to cell picker
mask must be on.  toggle labels with the L key (lower or uppercase,
doesn't matter).  uses the mask to construct labels via pymorph- so
requires the mask to be on

the display widget now uses 3 MPL objects for displaying--- imshow of
background, imshow of the mask and a list of all the text objects.

should be clean in terms of behavior, tested with average and single
frames.  re-draw is slow but it works and is accurate.