CellPicker: parameters, doLocalNMF, Info Panel + Edge bugs
sadly not split up into separate commits, but....

major thing here is returning (all lists):
   modes, thresholded modes, fit gaussians, this cell, is a cell

change way we calculate thresholded modes:  dilate then erode in
grayscale space, blur with a gaussian of 1px, then threshold using
otsu, follow with a size exclusion (75% of typical ROI volume).  if the
resulting thresholded mode is between 2 and 40% of all of the pixels,
AND the fit is near the middle of the ROI by morphology, we call it a
cell.

info panel changes to flip the modes int he info panel- the way we add
the axes flips them up/down.

fixed limit problems by returning the local limits of the NMF (why did
i calculate them twice?  and loop over the modes twice? a nightmare!)