Change to parseXSG
removing an accessor for literal pulses.  confused as to why this isn't
as deep as previously?

hope this doesn't break stuff for other people but i think i'm the only
one using literal pulses in the stimulator.

probably the better way to do this is to catch the TypeError that is
thrown...