ephusRoutines: removed 'parseAllXSGFiles' and added 'mergeXSGs', documented everything.
Lots of docstring changes.

wrote a function to merge two xsg files.  When merging, all non-numpy
fields are merged into a list.  Numpy fields are concatenated on a new
last axis, which is trials.  A new key is added to any dictionary that
is a combination of two xsgs, 'merged'.  Importantly, this function is
compatible with reduce().

Note that this function only works for xsgs of the same settings of the
ephus programs!  Merging behavior is undefined for acquisitions of
different length or different configurations of enabled ephus programs.

I removed parseAllXSGFiles() as 1) it was out of date and 2) this can
be easily and more flexibly accomplished using mergeXSGs() inside of a
reduce().  This is better--- no need to duplicate effort with two
import routines.